Refactor bsky.rs into modular components
Break down the large bsky.rs file into smaller modules:
- client_metadata.rs: OAuth client metadata endpoint
- auth.rs: Authorization flow and primary account selection
- callback.rs: OAuth callback handling
- profile.rs: User profile management
- tokens.rs: Token retrieval and revocation
- blobs.rs: Binary blob fetching
- utils.rs: Helper functions for OAuth operations

This modular structure improves code organization, maintainability,
and makes it easier to locate and update specific functionality.
